Use 16 colors

+++ hsluv-light.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 # pylint: disable=C0103,C0321,C0301
-"""Light scheme for Sigma-theme"""
+"""Light colors palette"""
 
 import os
 from hsluv import hsluv_to_rgb
@@ -9,10 +9,12 @@ from PIL import Image, ImageDraw
 
 # Preview
 xres = True
-# shot = False
+shot = True
 
-# Colors
+# Hue
 BA = 240
+
+# Sat
 SA = 30
 
 H00 = BA  ; S00 = 0 ; L00 = 100 ; COL00 = [] ; ZIP00 = [] #col00
@@ -27,11 +29,14 @@ H07 = BA  ; S07 = SA ; L07 = 65 ; COL07 = [] ; ZIP07 = [] #col07
 H08 = BA  ; S08 = SA ; L08 = 60 ; COL08 = [] ; ZIP08 = [] #col08
 H09 = BA  ; S09 = SA ; L09 = 55 ; COL09 = [] ; ZIP09 = [] #col09
 H10 = BA  ; S10 = SA ; L10 = 50 ; COL10 = [] ; ZIP10 = [] #col10
-H11 = BA  ; S11 = SA ; L11 = 35 ; COL11 = [] ; ZIP11 = [] #col11
-H12 = BA  ; S12 = SA ; L12 = 20 ; COL12 = [] ; ZIP12 = [] #col12
+H11 = BA  ; S11 = SA ; L11 = 45 ; COL11 = [] ; ZIP11 = [] #col11
+H12 = BA  ; S12 = SA ; L12 = 40 ; COL12 = [] ; ZIP12 = [] #col12
+H13 = BA  ; S13 = SA ; L13 = 35 ; COL13 = [] ; ZIP13 = [] #col13
+H14 = BA  ; S14 = SA ; L14 = 30 ; COL14 = [] ; ZIP14 = [] #col14
+H15 = BA  ; S15 = SA ; L15 = 25 ; COL15 = [] ; ZIP15 = [] #col15
+H16 = BA  ; S16 = SA ; L16 = 20 ; COL16 = [] ; ZIP16 = [] #col16
 
 # Conversion to RGB
-
 A00 = hsluv_to_rgb([H00, S00, L00])
 for i in A00: i = round(i*255); COL00.extend([i])
 
@@ -71,6 +76,17 @@ for i in A11: i = round(i*255); COL11.extend([i])
 A12 = hsluv_to_rgb([H12, S12, L12])
 for i in A12: i = round(i*255); COL12.extend([i])
 
+A13 = hsluv_to_rgb([H13, S13, L13])
+for i in A13: i = round(i*255); COL13.extend([i])
+
+A14 = hsluv_to_rgb([H14, S14, L14])
+for i in A14: i = round(i*255); COL14.extend([i])
+
+A15 = hsluv_to_rgb([H15, S15, L15])
+for i in A15: i = round(i*255); COL15.extend([i])
+
+A16 = hsluv_to_rgb([H16, S16, L16])
+for i in A16: i = round(i*255); COL16.extend([i])
 
 # Conversion to HEX
 for i in A00: i = round(i*255); ZIP00.extend([f'{i:02X}']); HEX00 = '#'+''.join(ZIP00)
@@ -86,13 +102,17 @@ for i in A09: i = round(i*255); ZIP09.extend([f'{i:02X}']); HEX09 = '#'+''.join(
 for i in A10: i = round(i*255); ZIP10.extend([f'{i:02X}']); HEX10 = '#'+''.join(ZIP10)
 for i in A11: i = round(i*255); ZIP11.extend([f'{i:02X}']); HEX11 = '#'+''.join(ZIP11)
 for i in A12: i = round(i*255); ZIP12.extend([f'{i:02X}']); HEX12 = '#'+''.join(ZIP12)
+for i in A13: i = round(i*255); ZIP13.extend([f'{i:02X}']); HEX13 = '#'+''.join(ZIP13)
+for i in A14: i = round(i*255); ZIP14.extend([f'{i:02X}']); HEX14 = '#'+''.join(ZIP14)
+for i in A15: i = round(i*255); ZIP15.extend([f'{i:02X}']); HEX15 = '#'+''.join(ZIP15)
+for i in A16: i = round(i*255); ZIP16.extend([f'{i:02X}']); HEX16 = '#'+''.join(ZIP16)
 
 # Define the individual width
-W = 120
-H = 70
+W = 200
+H = 40
 SEP = 10
 
-w, h = W*2+SEP*2, H*6+SEP*2
+w, h = W*1+SEP*2, H*16+SEP*2
 
 # Background
 BG = []
@@ -131,37 +151,58 @@ fig6 = ImageDraw.Draw(image)
 fig6.rectangle(SHAPE06, fill=(COL06[0], COL06[1], COL06[2]))
 
 # COL07
-SHAPE07 = [(W+SEP, H*0+SEP), (W+W+SEP, H*0+H+SEP)]
-fig7 = ImageDraw.Draw(image)
-fig7.rectangle(SHAPE07, fill=(COL07[0], COL07[1], COL07[2]))
+SHAPE07 = [(0+SEP, H*6+SEP), (W+SEP, H*6+H+SEP)]
+fig6 = ImageDraw.Draw(image)
+fig6.rectangle(SHAPE07, fill=(COL07[0], COL07[1], COL07[2]))
 
 # COL08
-SHAPE08 = [(W+SEP, H*1+SEP), (W+W+SEP, H*1+H+SEP)]
-fig8 = ImageDraw.Draw(image)
-fig8.rectangle(SHAPE08, fill=(COL08[0], COL08[1], COL08[2]))
+SHAPE08 = [(0+SEP, H*7+SEP), (W+SEP, H*7+H+SEP)]
+fig6 = ImageDraw.Draw(image)
+fig6.rectangle(SHAPE08, fill=(COL08[0], COL08[1], COL08[2]))
 
 # COL09
-SHAPE09 = [(W+SEP, H*2+SEP), (W+W+SEP, H*2+H+SEP)]
-fig9 = ImageDraw.Draw(image)
-fig9.rectangle(SHAPE09, fill=(COL09[0], COL09[1], COL09[2]))
+SHAPE09 = [(0+SEP, H*8+SEP), (W+SEP, H*8+H+SEP)]
+fig6 = ImageDraw.Draw(image)
+fig6.rectangle(SHAPE09, fill=(COL09[0], COL09[1], COL09[2]))
 
 # COL10
-SHAPE10 = [(W+SEP, H*3+SEP), (W+W+SEP, H*3+H+SEP)]
-fig10 = ImageDraw.Draw(image)
-fig10.rectangle(SHAPE10, fill=(COL10[0], COL10[1], COL10[2]))
+SHAPE10 = [(0+SEP, H*9+SEP), (W+SEP, H*9+H+SEP)]
+fig6 = ImageDraw.Draw(image)
+fig6.rectangle(SHAPE10, fill=(COL10[0], COL10[1], COL10[2]))
 
 # COL11
-SHAPE11 = [(W+SEP, H*4+SEP), (W+W+SEP, H*4+H+SEP)]
-fig11 = ImageDraw.Draw(image)
-fig11.rectangle(SHAPE11, fill=(COL11[0], COL11[1], COL11[2]))
+SHAPE11 = [(0+SEP, H*10+SEP), (W+SEP, H*10+H+SEP)]
+fig6 = ImageDraw.Draw(image)
+fig6.rectangle(SHAPE11, fill=(COL11[0], COL11[1], COL11[2]))
 
 # COL12
-SHAPE12 = [(W+SEP, H*5+SEP), (W+W+SEP, H*5+H+SEP)]
-fig12 = ImageDraw.Draw(image)
-fig12.rectangle(SHAPE12, fill=(COL12[0], COL12[1], COL12[2]))
+SHAPE12 = [(0+SEP, H*11+SEP), (W+SEP, H*11+H+SEP)]
+fig6 = ImageDraw.Draw(image)
+fig6.rectangle(SHAPE12, fill=(COL12[0], COL12[1], COL12[2]))
+
+# COL13
+SHAPE13 = [(0+SEP, H*12+SEP), (W+SEP, H*12+H+SEP)]
+fig6 = ImageDraw.Draw(image)
+fig6.rectangle(SHAPE13, fill=(COL13[0], COL13[1], COL13[2]))
+
+# COL14
+SHAPE14 = [(0+SEP, H*13+SEP), (W+SEP, H*13+H+SEP)]
+fig6 = ImageDraw.Draw(image)
+fig6.rectangle(SHAPE14, fill=(COL14[0], COL14[1], COL14[2]))
+
+# COL15
+SHAPE15 = [(0+SEP, H*14+SEP), (W+SEP, H*14+H+SEP)]
+fig6 = ImageDraw.Draw(image)
+fig6.rectangle(SHAPE15, fill=(COL15[0], COL15[1], COL15[2]))
+
+# COL16
+SHAPE16 = [(0+SEP, H*15+SEP), (W+SEP, H*15+H+SEP)]
+fig6 = ImageDraw.Draw(image)
+fig6.rectangle(SHAPE16, fill=(COL16[0], COL16[1], COL16[2]))
 
 # Output
 if xres is True:
+    print("col00", HEX00)
     print("col01", HEX01)
     print("col02", HEX02)
     print("col03", HEX03)
@@ -174,6 +215,10 @@ if xres is True:
     print("col10", HEX10)
     print("col11", HEX11)
     print("col12", HEX12)
+    print("col13", HEX13)
+    print("col14", HEX14)
+    print("col15", HEX15)
+    print("col16", HEX16)
 
 # Generate the image
 IMAGE='light.png'
@@ -183,7 +228,8 @@ PS=str("ps -ef | grep -v grep | grep 'mupdf' ")
 FL=str("| grep -o '" + IMAGE + "' > /dev/null")
 CHECK=os.system(PS+FL)
 
-if CHECK == 256:
-    os.system("mupdf '" + IMAGE + "' 2>/dev/null &")
-else:
-    os.system("pkill -HUP mupdf 2>/dev/null")
+if shot is True:
+    if CHECK == 256:
+        os.system("mupdf '" + IMAGE + "' 2>/dev/null &")
+    else:
+        os.system("pkill -HUP mupdf 2>/dev/null")
